This PR was squashed before being merged into the 3.1-dev branch (closes #17598).

Discussion
----------

Improved FileResource and DirectoryResource

| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | yes
| BC breaks?    | yes?
| Deprecations? | no
| Tests pass?   | yes
| Fixed tickets | #15946
| License       | MIT
| Doc PR        | -

Commits
-------

3e762a6 Improved FileResource and DirectoryResource

+++ src/Symfony/Component/Config/Tests/Resource/FileResourceTest.php
@@ -29,6 +29,10 @@ class FileResourceTest extends \PHPUnit_Framework_TestCase
 
     protected function tearDown()
     {
+        if (!file_exists($this->file)) {
+            return;
+        }
+
         unlink($this->file);
     }
 
@@ -42,14 +46,27 @@ class FileResourceTest extends \PHPUnit_Framework_TestCase
         $this->assertSame(realpath($this->file), (string) $this->resource);
     }
 
+    /**
+     * @expectedException \InvalidArgumentException
+     * @expectedExceptionMessageRegExp /The file ".*" does not exist./
+     */
+    public function testResourceDoesNotExist()
+    {
+        $resource = new FileResource('/____foo/foobar'.mt_rand(1, 999999));
+    }
+
     public function testIsFresh()
     {
         $this->assertTrue($this->resource->isFresh($this->time), '->isFresh() returns true if the resource has not changed in same second');
         $this->assertTrue($this->resource->isFresh($this->time + 10), '->isFresh() returns true if the resource has not changed');
         $this->assertFalse($this->resource->isFresh($this->time - 86400), '->isFresh() returns false if the resource has been updated');
+    }
 
-        $resource = new FileResource('/____foo/foobar'.mt_rand(1, 999999));
-        $this->assertFalse($resource->isFresh($this->time), '->isFresh() returns false if the resource does not exist');
+    public function testIsFreshForDeletedResources()
+    {
+        unlink($this->file);
+
+        $this->assertFalse($this->resource->isFresh($this->time), '->isFresh() returns false if the resource does not exist');
     }
 
     public function testSerializeUnserialize()
